Add : mock data 이용하 여 Feed component rendering
