+++ src/components/Hyojeong/Feed.js
@@ -1,9 +1,9 @@
 import React, { useState, useRef } from 'react';
-import FeedTop from './FeedTop';
 import CommentList from './CommentList';
 
-const Feed = () => {
+const Feed = ({ feed }) => {
   let commentId = useRef(4);
+  const commentForm = useRef();
   const [comments, setComments] = useState([]);
   const [comment, setComment] = useState({
     id: commentId.current,
@@ -11,8 +11,6 @@ const Feed = () => {
     content: '',
     isLiked: false,
   });
-
-  const commentForm = useRef();
 
   // comment handling - set comments in an array when they are submitted
   const submitComment = e => {
@@ -47,11 +45,41 @@ const Feed = () => {
     );
   };
 
+  //paint heart btn or bookmark btn in the feed-bar
+  const paintBtn = e => {
+    e.target.classList.toggle('fas');
+  };
+
   return (
     <div className="Feed">
-      <FeedTop />
+      <div className="feed-bar">
+        <div className="profile-box">
+          <img alt="profile" className="profile" src={feed.userProfileImg} />
+          <strong className="profile-id">{feed.userName}</strong>
+        </div>
+
+        <i className="fas fa-ellipsis-h" />
+      </div>
+      <div className="feed-photo">
+        <img alt="feed" src={feed.thumbnail} />
+      </div>
+      <div className="feed-btns">
+        <button className="feed-heart-btn" onClick={paintBtn}>
+          <i className="far fa-heart" />
+        </button>
+        <button>
+          <i className="far fa-comment" />
+        </button>
+        <button>
+          <i className="fas fa-external-link-alt" />
+        </button>
+        <button className="absolute" onClick={paintBtn}>
+          <i className="far fa-bookmark" />
+        </button>
+      </div>
+
       <div className="user-info feed-description">
-        <strong className="profile-id">canon_mj</strong> 날 좋은 하루💙
+        <strong className="profile-id">{feed.userName}</strong> {feed.content}
         <br />
         <div className="scroll">
           <CommentList
