+++ src/pages/Hyojeong/Main/Main.scss
@@ -49,53 +49,50 @@
       border-radius: $border-radius;
       background-color: white;
 
-      /* FeedTop component */
-      .FeedTop {
-        .feed-bar {
-          @include flex-space-between;
-          align-items: center;
-          padding: 15px 30px;
+      .feed-bar {
+        @include flex-space-between;
+        align-items: center;
+        padding: 15px 30px;
 
-          .profile-box {
-            display: flex;
-            align-items: center;
-          }
-
-          i {
-            font-size: 20px;
-            cursor: pointer;
-          }
-        }
-        .feed-photo {
-          width: 100%;
-          height: 630px;
-
-          img {
-            width: inherit;
-            height: 100%;
-          }
-        }
-        .feed-btns {
-          position: relative;
-          flex: 0.7;
+        .profile-box {
           display: flex;
-          height: 50px;
-          padding: 10px;
+          align-items: center;
+        }
 
-          button {
-            margin: 0 2px;
-            border: 0;
-            background-color: transparent;
-            color: #414141;
-            font-size: 30px;
-          }
+        i {
+          font-size: 20px;
+          cursor: pointer;
+        }
+      }
+      .feed-photo {
+        width: 100%;
+        height: 630px;
 
-          .absolute {
-            position: absolute;
-            right: 10px;
-            bottom: 50%;
-            transform: translateY(50%);
-          }
+        img {
+          width: inherit;
+          height: 100%;
+        }
+      }
+      .feed-btns {
+        position: relative;
+        flex: 0.7;
+        display: flex;
+        height: 50px;
+        padding: 10px;
+
+        button {
+          margin: 0 2px;
+          border: 0;
+          background-color: transparent;
+          color: #414141;
+          font-size: 30px;
+        }
+
+        .absolute {
+          position: absolute;
+          right: 10px;
+          bottom: 50%;
+          transform: translateY(50%);
         }
       }
